+++ packages/textlab/test/index.test.ts
@@ -1,5 +1,0 @@
-import { packageName } from "../src/index.ts";
-
-const expectedPackageName: typeof packageName = "@ismail-elkorchi/textlab";
-
-void expectedPackageName;